Feature: Schema
This changeset adds the central schema for the *COVID Impact Map*
application. This schema is designed to be used with the database,
API layer, and eventually the frontend, to render the map UI and
detail screens.

First up, we need a datamodel for healthcare facilities that mirrors
the one established by [this Google sheet](https://docs.google.com/spreadsheets/d/1Qm11H2R3kom5cRqoPcC3ObnVWzAUp_yPSfO57tPQdUo/edit#gid=0).
Next, we need a reporting model that is flexible enough to allow
easy crowdsourced reporting, but specific enough to have meaning
when we query it later on.

So here it is:
- [x] Add `Facility` schema, with contact info, location info, and
      basic categorical structures/properties
- [x] Beginnings of reporting schema
- [ ] Fleshed out reporting schema
- [ ] Structure for summed-up facility stats
- [ ] Main API interface
  - [x] Healthcheck method
  - [x] Facilities list method with query parameters
  - [ ] Ability to fetch stats for nearby facilities
    - [ ] Input from address
    - [ ] Input from lat/long